Feat: 사이드바 채팅방 메뉴에 보관, 삭제, 이름 변경 버튼 구조화 진행 (컴포넌트 분리)
