+++ client/src/widgets/layout/sidebar/components/ChatRoomItemMenu.tsx
@@ -1,59 +1,19 @@
-import { Dispatch, SetStateAction, useState } from "react";
-import { useLocation, useNavigate } from "react-router-dom";
-import { useShallow } from "zustand/shallow";
-import { useDeleteChatRoom, useUpdateChatRoomTitle } from "@/widgets/layout/sidebar/queries";
+import { Dispatch, SetStateAction } from "react";
+import { useLocation } from "react-router-dom";
 import { Separator } from "@/shared/ui/separator";
-import { DropdownMenuItem, DropdownMenuContent } from "@/shared/ui/dropdown-menu";
-import { ConfirmDialog, EditDialog } from "@/shared/components";
-import useChatRoomStore from "@/shared/stores/useChatRoomStore";
-import { Pencil, Trash2 } from "lucide-react";
+import { DropdownMenuContent } from "@/shared/ui/dropdown-menu";
+import { ChatRoomItemArchiveButton, ChatRoomItemDeleteButton, ChatRoomItemRenameButton } from "./menu-buttons";
 
 interface ChatRoomItemMenuPropsType {
   roomId: string;
   roomTitle: string;
   setHoveredRoom: Dispatch<SetStateAction<string | null>>;
 }
+
 const ChatRoomItemMenu = ({ roomId, roomTitle, setHoveredRoom }: ChatRoomItemMenuPropsType) => {
-  const navigate = useNavigate();
   const { pathname } = useLocation();
-  const [showConfirmDialog, setShowConfirmDialog] = useState(false);
-  const [showEditDialog, setShowEditDialog] = useState(false);
-
-  const [updateChatRoomTitle, deleteChatRoom] = useChatRoomStore(
-    useShallow((state) => [state.updateChatRoomTitle, state.deleteChatRoom]),
-  );
-
-  const updateChatRoomTitleMutation = useUpdateChatRoomTitle();
-  const deleteChatRoomMutation = useDeleteChatRoom();
 
   const isNowChatRoom = pathname === "/chat/" + roomId;
-  const alertMessage = isNowChatRoom ? "현재 채팅방이 삭제되고 메인페이지로 이동합니다." : "해당 채팅방이 삭제됩니다.";
-
-  // 채팅방 삭제
-  const handleDelete = () => {
-    deleteChatRoomMutation.mutate(
-      { params: { roomId } },
-      {
-        onSuccess: () => {
-          deleteChatRoom(roomId);
-          if (isNowChatRoom) navigate("/");
-        },
-      },
-    );
-  };
-
-  // 채팅방 이름 변경
-  const handleRename = (newTitle: string) => {
-    updateChatRoomTitleMutation.mutate(
-      { roomId, newTitle },
-      {
-        onSuccess: () => {
-          updateChatRoomTitle(roomId, newTitle);
-        },
-      },
-    );
-    setHoveredRoom(null);
-  };
 
   return (
     <>
@@ -63,42 +23,12 @@ const ChatRoomItemMenu = ({ roomId, roomTitle, setHoveredRoom }: ChatRoomItemMen
         sideOffset={-5}
         alignOffset={10}
       >
-        <DropdownMenuItem
-          className="dark:hover:bg-accent flex cursor-pointer items-center gap-3 rounded-md p-2 text-sm"
-          onClick={() => setShowEditDialog(true)}
-        >
-          <Pencil />
-          <span>이름 변경</span>
-        </DropdownMenuItem>
+        <ChatRoomItemRenameButton roomId={roomId} roomTitle={roomTitle} setHoveredRoom={setHoveredRoom} />
+        <ChatRoomItemArchiveButton roomId={roomId} isNowChatRoom={isNowChatRoom} />
+
         <Separator />
-        <DropdownMenuItem
-          className="dark:hover:bg-accent flex cursor-pointer items-center gap-3 rounded-md p-2 text-sm text-red-400 hover:text-red-400"
-          onClick={() => setShowConfirmDialog(true)}
-        >
-          <Trash2 />
-          <span>삭제</span>
-        </DropdownMenuItem>
+        <ChatRoomItemDeleteButton roomId={roomId} isNowChatRoom={isNowChatRoom} />
       </DropdownMenuContent>
-
-      {/* 삭제 Alert */}
-      <ConfirmDialog
-        title="채팅방을 삭제하시겠습니까?"
-        description={alertMessage}
-        open={showConfirmDialog}
-        onOpenChange={setShowConfirmDialog}
-        onConfirm={handleDelete}
-        confirmText="삭제"
-        confirmColor="bg-red-500 hover:bg-red-600 dark:bg-red-500 hover:dark:bg-red-600"
-      />
-
-      {/* 이름 변경 Alert */}
-      <EditDialog
-        title="채팅방 이름 변경"
-        open={showEditDialog}
-        inputValue={roomTitle}
-        onOpenChange={setShowEditDialog}
-        onConfirm={handleRename}
-      />
     </>
   );
 };
