+++ client/src/widgets/layout/sidebar/components/menu-buttons/index.ts
@@ -1,0 +1,5 @@
+import ChatRoomItemArchiveButton from "./ChatRoomItemArchiveButton";
+import ChatRoomItemDeleteButton from "./ChatRoomItemDeleteButton";
+import ChatRoomItemRenameButton from "./ChatRoomItemRenameButton";
+
+export { ChatRoomItemArchiveButton, ChatRoomItemDeleteButton, ChatRoomItemRenameButton };
